Improve predefined pack/unpack performance using mvapich methods.
For reference implementation, see:
https://github.com/pmodels/mpich/blob/9ab5fd06af2a648bf24214f0d9cff0ee77ee3e7d/src/mpi/datatype/veccpy.h

The overhead of memcpy() was causing noticable slowdown in the
predefined pack/unpack routines. To fix that, this PR implements a
manual copy for blocklengths of <= 8, an idea borrowed from
mpich above.

It may also be beneficial to perform a manual copy for larger blocklengths.
However, more data will have to be gathered to see if a similar
implementation would be better over the current memcpy() approach.